tests/wpt/tests/scroll-animations/animation-trigger/parsing/timeline-trigger-behavior-valid.tentative.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: parsing timeline-trigger-behavior with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-animations-2/#animation-trigger">
<meta name="assert" content="animation-composition supports the full grammar '<single-timeline-trigger-behavior> #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("timeline-trigger-behavior", "once");
test_valid_value("timeline-trigger-behavior", "repeat");
test_valid_value("timeline-trigger-behavior", "alternate");
test_valid_value("timeline-trigger-behavior", "state");
test_valid_value("timeline-trigger-behavior", "once, repeat");
test_valid_value("timeline-trigger-behavior", "once, repeat, alternate");
test_valid_value("timeline-trigger-behavior", "once, repeat, alternate, state");

test_invalid_value("timeline-trigger-behavior", "oncerepeat");
test_invalid_value("timeline-trigger-behavior", "repeatalternate");
test_invalid_value("timeline-trigger-behavior", "once, repeatalternate");
test_invalid_value("timeline-trigger-behavior", "junk");
test_invalid_value("timeline-trigger-behavior", "enco, taeper, etanretla, etats");
</script>
</body>
</html>
